For clarity (and mostly for fun), moving from CamelCase to kebab-case to align with actual tag names. Just easier to reason about, for now.
