+++ demo/vercel/src/lib/features-info.svelte
