+++ demo/vercel/src/lib/tabs-demo.svelte
@@ -1,6 +1,6 @@
 <script>
 	import { TabsWrapper, TabList, TabPanel, TabButton } from './tabs';
-	import Counter from './Counter.svelte';
+	import Counter from './counter-tag.svelte';
 </script>
 
 <TabsWrapper>
@@ -19,23 +19,23 @@
 
 			<TabPanel>
 				<h2>First nested panel</h2>
-				<Counter count="1" award="5"/>
+				<Counter count={1} award={5} />
 			</TabPanel>
 
 			<TabPanel>
 				<h2>Second nested panel</h2>
-				<Counter count="2" award="5"/>
+				<Counter count={2} award={5} />
 			</TabPanel>
 		</TabsWrapper>
 	</TabPanel>
 
 	<TabPanel>
 		<h2>Second panel</h2>
-		<Counter count="3" award="5"/>
+		<Counter count={3} award={5} />
 	</TabPanel>
 
 	<TabPanel>
 		<h2>Third panel</h2>
-		<Counter count="4" award="5"/>
+		<Counter count={4} award={5} />
 	</TabPanel>
 </TabsWrapper>
