+++ demo/vercel/src/lib/tabs-info.svelte
