+++ demo/vercel/src/lib/tabs/tab-panel.svelte
@@ -1,6 +1,6 @@
 <script>
 	import { getContext } from 'svelte';
-	import { TABS } from './TabsWrapper.svelte';
+	import { TABS } from './tabs-wrapper.svelte';
 
 	const panel = {};
 	const testContext = getContext(TABS);
